Move slirp4netns functions into an extra file
This should make maintenance easier.
